update: separate print funcs for recv
